Simplify XML parsing
Instead of parsing XML in a streaming fashion, this parses it all at
once.  This requires storing some of the data that's in the XML longer
(in the data structure) rather than ephemerally.

Instead of threading the encryption key for the passwords through by
overriding the UnmarshalXML method, this avoids the need by walking the
data structure after the fact.

Major drawback of this approach: we lose some of the validation built
into the XML unmarshaling, specifically around History.